### 2021-10-01 #3792 @whedon
[PRE REVIEW]: gym-saturation: an OpenAI Gym environment for saturation provers
**Submitting author:** @inpefess (<a href="http://orcid.org/0000-0002-1291-9896">Boris Shminke</a>)
**Repository:** <a href="https://github.com/inpefess/gym-saturation" target ="_blank">https://github.com/inpefess/gym-saturation</a>
**Version:** v0.1.0
