### 2021-10-01 #3792 @danielskatz
@inpefess - while we look for an editor, you could work on the possibly missing DOIs that whedon suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references from branch joss-paper` to check again, and the command `@whedon generate pdf from branch joss-paper` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
