### 2021-10-05 d22ed91
regenerate doc index from README
