+++ doc/source/index.rst
@@ -13,47 +13,70 @@
   See the License for the specific language governing permissions and
   limitations under the License.
 
-==========================================
- Welcome to Saturation Gym documentation!
-==========================================
-
-``gym-saturation`` is an `OpenAI Gym`_ environment for reinforcement learning (RL) agents capable of proving theorems. Currently, only theorems in `CNF`_ sublanguage of `TPTP`_ are supported. ``gym-saturation`` implements the 'given clause' algorithm (similar to one used in `Vampire`_ and `E Prover`_). Although, being written in Python, ``gym-saturation`` is closer to `PyRes`_. In contrast to monolithic architecture of a typical ATP, ``gym-saturation`` gives different agents opportunities to select clauses themselves and train from their experience. Combined with a particular agent, ``gym-saturation`` can work as an Automated Theorem Prover (ATP).
-
-``gym-saturation`` can be interesting for RL practicioners willing to apply their experience to theorem proving without coding all the logic-related stuff themselves. It also can be useful for automated deduction researchers who want to create an RL-empowered ATP.
-
-How to install
+gym-saturation
 ==============
 
-The best way to install this package is to use ``pip``::
+``gym-saturation`` is an `OpenAI Gym <https://gym.openai.com/>`__
+environment for reinforcement learning (RL) agents capable of proving
+theorems. Currently, only theorems in CNF sublanguage of
+`TPTP <http://tptp.org>`__ are supported. ``gym-saturation`` implements
+the ‘given clause’ algorithm (similar to one used in
+`Vampire <https://github.com/vprover/vampire>`__ and `E
+Prover <https://github.com/eprover/eprover>`__). Although, being written
+in Python, ``gym-saturation`` is closer to
+`PyRes <https://github.com/eprover/PyRes>`__. In contrast to monolithic
+architecture of a typical ATP, ``gym-saturation`` gives different agents
+opportunities to select clauses themselves and train from their
+experience. Combined with a particular agent, ``gym-saturation`` can
+work as an Automated Theorem Prover (ATP).
 
-    pip install gym-saturation
-    
+``gym-saturation`` can be interesting for RL practicioners willing to
+apply their experience to theorem proving without coding all the
+logic-related stuff themselves. It also can be useful for automated
+deduction researchers who want to create an RL-empowered ATP.
+
+How to Install
+==============
+
+The best way to install this package is to use ``pip``:
+
+.. code:: sh
+
+   pip install gym-saturation
+
 How to use
 ==========
 
-See `the notebook`_ for more information.
+See `the
+notebook <https://github.com/inpefess/gym-saturation/blob/master/examples/example.ipynb>`__
+for more information.
 
-How to contribute
+How to Contribute
 =================
 
-`Pull requests`_ are welcome. To start::
+`Pull requests <https://github.com/inpefess/gym-saturation/pulls>`__ are
+welcome. To start:
 
-    git clone https://github.com/inpefess/gym-saturation
-    cd gym-saturation
-    # activate python virtual environment with Python 3.6+
-    pip install -U pip
-    pip install -U setuptools wheel poetry
-    poetry install
-    # recommended but not necessary
-    pre-commit install
+.. code:: sh
 
-To check the code quality before creating a pull request, one might run the script ``show_report.sh``. It locally does nearly the same as the CI pipeline after the PR is created.
+   git clone https://github.com/inpefess/gym-saturation
+   cd gym-saturation
+   # activate python virtual environment with Python 3.6+
+   pip install -U pip
+   pip install -U setuptools wheel poetry
+   poetry install
[…]
