### 2021-10-05 7fd7517
remove redundant type definition
