### 2021-10-06 2d5846c
fix bugs in docstring and style inconsistensies
