### 2021-10-06 #3792 @VivianePons
@inpefess do you have some suggestion of reviewers? especially based on our list here: https://bit.ly/joss-reviewers (But you can also suggest people outside this list)
Please, if they are on github don't @ them directly, just copy their handle to avoid them getting extra notifications
