### 2021-10-25 #3849 @whedon
[REVIEW]: gym-saturation: an OpenAI Gym environment for saturation provers
**Submitting author:** @inpefess (<a href="http://orcid.org/0000-0002-1291-9896">Boris Shminke</a>)
**Repository:** https://github.com/inpefess/gym-saturation
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.9
