+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "gym-saturation"
-version = "0.1.2"
+version = "0.1.3"
 description = "An OpenAI Gym environment for saturation provers"
 authors = ["Boris Shminke <boris@shminke.ml>"]
 license = "Apache-2.0"
@@ -10,7 +10,7 @@ classifiers=[
 	"Programming Language :: Python :: 3.6",
 	"Programming Language :: Python :: 3.7",
 	"Programming Language :: Python :: 3.8",
-	"Programming Language :: Python :: 3.9",	
+	"Programming Language :: Python :: 3.9",
 	"License :: OSI Approved :: Apache Software License",
 	"Operating System :: OS Independent",
 	"Intended Audience :: Science/Research",
@@ -23,14 +23,14 @@ classifiers=[
 include = ["gym_saturation/py.typed"]
 
 [tool.poetry.dependencies]
-python = ">= 3.6.1, < 4.0"
+python = ">= 3.6.1, < 3.10"
 lark-parser = "*"
 gym = "*"
 importlib_resources = {version = "*", markers = "python_version < \"3.9\" and python_version >= \"3.7\""}
 dataclasses = {version = "*", markers = "python_version < \"3.7\""}
 
 [tool.poetry.dev-dependencies]
-jedi = "0.17.2"
+jedi = "*"
 rope = "*"
 black = "*"
 jupyterlab = "*"
@@ -55,5 +55,5 @@ profile = "black"
 src_paths = ["gym_saturation"]
 
 [build-system]
-requires = ["poetry-core>=1.1.7"]
+requires = ["poetry-core>=1.1.11"]
 build-backend = "poetry.core.masonry.api"
