+++ .circleci/config.yml
@@ -1,23 +1,19 @@
 version: 2.1
 
-orbs:
-  python: circleci/python@0.2.1
-
 jobs:
   build-and-test:
     docker:
-      - image: circleci/python:3.6.14
+      - image: cimg/python:3.9.7
     working_directory: ~/repo
     steps:
       - checkout
       - restore_cache:
           keys:
             - v1-dependencies-{{ checksum "poetry.lock" }}
-            - v1-dependencies-
       - run:
           name: install dependencies
           command: |
-            python3 -m venv venv
+            python -m venv venv
             . venv/bin/activate
             pip install -U pip
             pip install -U wheel setuptools poetry
