+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "gym-saturation"
-version = "0.1.3"
+version = "0.1.4"
 description = "An OpenAI Gym environment for saturation provers"
 authors = ["Boris Shminke <boris@shminke.ml>"]
 license = "Apache-2.0"
