### 2021-11-05 #3849 @inpefess
@ian-bunner and @lutzhamel I updated the code (not the paper) today (added Dockerfile for Binder mostly). Please use the latest package version, since the documentation was updated too.
