+++ gym_saturation/envs/saturation_env.py
@@ -50,9 +50,9 @@ class SaturationEnv(Env):
     ...     from importlib_resources import files
     >>> tptp_folder = files("gym_saturation").joinpath("resources/TPTP-mock")
     >>> from glob import glob
-    >>> problem_list = glob(
+    >>> problem_list = sorted(glob(
     ...     os.path.join(tptp_folder, "Problems", "*", "*1-1.p")
-    ... )
+    ... ))
     >>> env = SaturationEnv(3, problem_list)
 
     there is nothing non-deterministic here, but the seed can be set
