+++ doc/source/testing-an-agent.rst
@@ -14,19 +14,19 @@
   limitations under the License.
 
 #################
-Testing a Policy
+Testing an Agent
 #################
 
-Suppose you already have a trained policy implemented. Then you can use `a policy testing script`_ in parallel like that::
+Suppose you already have a trained agent implemented. Then you can use `an agent testing script`_ in parallel like that::
 
-  find $TPTP_HOME/Problems/*/*-*.p | parallel --bar --jobs 80% --timeout 30000% python policy_testing.py --problem_file {} --output_folder TPTP_CNF --step_limit 20
+  find $TPTP_HOME/Problems/*/*-*.p | parallel --bar --jobs 80% --timeout 30000% python agent_testing.py --problem_file {} --output_folder TPTP_CNF --step_limit 20
 
-You can write your own policy testing script based on ``policy_testing.py`` by calling ``episode`` function with your policy as an argument.
+You can write your own agent testing script based on ``agent_testing.py`` by calling ``episode`` function with your agent as an argument.
 
 After processing problems, you can get a report about its performance::
 
   import os
-  from gym_saturation.policy_testing import policy_testing_report
+  from gym_saturation.agent_testing import agent_testing_report
   from glob import glob
   import sys
 
@@ -34,18 +34,18 @@ After processing problems, you can get a report about its performance::
   problem_list = glob(os.path.join(
       os.environ["TPTP_HOME"], "Problems", "*", "*-*.p")
   )
-  report = policy_testing_report(problem_list, "TPTP_CNF_20")
+  report = agent_testing_report(problem_list, "TPTP_CNF_20")
 
 Applying different policies ``gym-saturation`` leads to the following results on all CNF problems from TPTP-v7.5.0 with the step limit set to 20:
 
 +--------+----------------+-------------+--------------------+-------+
-| policy | total problems | proof found | step limit reached | error |
+| agent  | total problems | proof found | step limit reached | error |
 +========+================+=============+====================+=======+
 | size   | 8257           | 2245        | 5889               | 123   |
 +--------+----------------+-------------+--------------------+-------+
 | age    | 8257           | 234         | 7884               | 139   |
 +--------+----------------+-------------+--------------------+-------+
 
-Can your policy do better?
+Can your agent do better?
 
-.. _a policy testing script: https://github.com/inpefess/gym-saturation/blob/master/gym_saturation/policy_testing.py
+.. _an agent testing script: https://github.com/inpefess/gym-saturation/blob/master/gym_saturation/agent_testing.py
