+++ gym_saturation/agent_testing.py
@@ -20,7 +20,9 @@ import sys
 from argparse import ArgumentParser, Namespace
 from dataclasses import dataclass
 from operator import itemgetter
-from typing import Any, Callable, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Optional, Tuple
+
+import gym
 
 from gym_saturation.envs import SaturationEnv
 from gym_saturation.grammar import Clause
@@ -28,27 +30,47 @@ from gym_saturation.logic_ops.utils import clause_length
 from gym_saturation.parsing.json_grammar import dict_to_clause
 
 
+# pylint: disable=too-few-public-methods
+class BaseAgent:
+    """ a basic RL agent """
+
+    def get_action(
+        self,
+        observation: List[Dict[str, Any]],
+        reward: float,
+        info: Dict[str, Any],
+    ) -> int:
+        """
+        :param observation: an observation returned by the environment after
+            the latest action (or it's initial state)
+        :param reward: reward from the previous step
+            (use zero for the first step)
+        :param info: info dictionary returned by the environment
+        :returns: an action
+        """
+
+
 @dataclass
 class Transition:
     """
-    an object describing environment's state before the agent's action, the
-    action itself and its results
+    an object describing environment's and agent's states
+    before the agent's action, the action itself and its results
     """
 
-    state: List[Dict[str, Any]]
+    env_state: List[Dict[str, Any]]
+    agent_state: Any
     action: int
-    agent_info: Dict[str, Any]
-    next_state: List[Dict[str, Any]]
+    observation: List[Dict[str, Any]]
     reward: float
     done: bool
-    env_info: Dict[str, Any]
+    info: Dict[str, Any]
 
 
 def save_final_state(
     problem_filename: str, output_folder: str, episode_memory: List[Transition]
 ) -> None:
     """
-    save the final state of solving a TPTP problem
+    save the final environment state of solving a TPTP problem
 
     :param problem_filename: a full path to the problem file
     :param output_folder: where to write subfolders and JSON files
@@ -68,81 +90,69 @@ def save_final_state(
         "w",
         encoding="utf-8",
     ) as data_file:
-        data_file.write(json.dumps(episode_memory[-1].next_state))
+        data_file.write(json.dumps(episode_memory[-1].observation))
 
 
-# pylint: disable=unused-argument
-def size_agent(
-    state: list, agent_info: Dict[str, Any], env_info: Dict[str, Any]
-) -> Tuple[int, Dict[str, Any]]:
-    """
-    an example of an implemented agent
+class SizeAgent(BaseAgent):
+    """ agent which selects the shortest clause """
 
-    :param state: a list of clauses
-    :param agent_info: added for compatibility purpose
-    :param env_info: added for compatibility purpose
-    :returns: the index of the clause with minimal length, empty info dict
-    """
-    return (
-        min(
+    def get_action(
+        self,
+        observation: List[Dict[str, Any]],
+        reward: float,
+        info: Dict[str, Any],
+    ) -> int:
+        return min(
             [
                 (i, clause_length(dict_to_clause(clause)))
-                for i, clause in enumerate(state)
+                for i, clause in enumerate(observation)
                 if not clause["processed"]
             ],
             key=itemgetter(1),
-        )[0],
-        {},
-    )
+        )[0]
 
 
-# pylint: disable=unused-argument
-def age_agent(
-    state: list, agent_info: Dict[str, Any], env_info: Dict[str, Any]
-) -> Tuple[int, Dict[str, Any]]:
-    """
-    another example of an implemented agent
+class AgeAgent(BaseAgent):
+    """ agent which selects the oldest clause """
 
[…]
