+++ gym_saturation/envs/saturation_env.py
@@ -44,7 +44,7 @@ class SaturationEnv(Env):
     saturation algorithm defined in a Reiforcement Learning friendly way
 
     >>> import sys
-    >>> if sys.version_info.major == 3 and sys.version_info.minor == 9:
+    >>> if sys.version_info.major == 3 and sys.version_info.minor >= 9:
     ...     from importlib.resources import files
     ... else:
     ...     from importlib_resources import files
