### 2021-11-16 #3849 @lutzhamel
@VivianePons @inpefess Here are some initial comments on the software and the paper,

1. I could not find any unit tests.  If there are any they don't seem to accessible to the user.
2. I think there are too many acronyms in the introductory paragraph which are not being properly defined before use, e.g CNF, TPTP, ATP
3. The sentence "Although, being written in Python, gym-saturation is closer to PyRes (Schulz & Pease, 2020)."
makes no sense to me since both packages are written in Python. Perhaps this sentence and the one before it need to slightly rewritten.
4. Would it make sense to have a simple example of how to use the package in the paper?
