### 2021-11-16 2f51022
fewer acronyms and a bit better wording
