+++ joss-paper/paper.md
@@ -18,7 +18,7 @@ bibliography: paper.bib
 # Summary
 ---
 
-`gym-saturation` is an OpenAI Gym [@DBLP:journals/corr/BrockmanCPSSTZ16] environment for reinforcement learning (RL) agents capable of proving theorems. Currently, only theorems in the CNF sublanguage of TPTP [@Sut17] are supported. `gym-saturation` implements the 'given clause' algorithm (similar to the one used in Vampire [@DBLP:conf/cav/KovacsV13] and E Prover [@DBLP:conf/cade/0001CV19]). Although, being written in Python, `gym-saturation` is closer to PyRes [@DBLP:conf/cade/0001P20]. In contrast to the monolithic architecture of a typical ATP, `gym-saturation` gives different agents opportunities to select clauses themselves and train from their experience. Combined with a particular agent, `gym-saturation` can work as an Automated Theorem Prover (ATP). Even with a non trained agent based on the clause size heuristic, `gym-saturation` can find refutations for 2245 (of 8257) CNF problems from TPTP v7.5.0.
+`gym-saturation` is an OpenAI Gym [@DBLP:journals/corr/BrockmanCPSSTZ16] environment for reinforcement learning (RL) agents capable of proving theorems. Currently, only theorems written in TPTP library [@Sut17] formal language in clausal normal form (CNF) are supported. `gym-saturation` implements the 'given clause' algorithm (similar to the one used in Vampire [@DBLP:conf/cav/KovacsV13] and E Prover [@DBLP:conf/cade/0001CV19]). Being written in Python, `gym-saturation` was inspired by PyRes [@DBLP:conf/cade/0001P20]. In contrast to the monolithic architecture of a typical Automated Theorem Prover (ATP), `gym-saturation` gives different agents opportunities to select clauses themselves and train from their experience. Combined with a particular agent, `gym-saturation` can work as an ATP. Even with a non trained agent based on the clause size heuristic, `gym-saturation` can find refutations for 2245 (of 8257) CNF problems from TPTP v7.5.0.
 
 # Statement of need
 
