### 2021-11-16 #3849 @inpefess
@lutzhamel thank you for your comments:

>     1. I could not find any unit tests.  If there are any they don't seem to accessible to the user.

This package uses [doctest](https://docs.python.org/3/library/doctest.html) and is 100% covered. The tests are accessible to a user as part of docstrings of functions and classes, like [here](https://gym-saturation.readthedocs.io/en/latest/package-documentation.html#gym_saturation.logic_ops.substitution.Substitution)

>     2. I think there are too many acronyms in the introductory paragraph which are not being properly defined before use, e.g CNF, TPTP, ATP

Thank you for spotting that! Unfortunately, for me, they became too recognisable through time. I added full wording, except TPTP, since it uses the acronym on its [main page](http://tptp.org/) caption.

>     3. The sentence "Although, being written in Python, gym-saturation is closer to PyRes (Schulz & Pease, 2020)."
>        makes no sense to me since both packages are written in Python. Perhaps this sentence and the one before it need to slightly rewritten.

Thanks, that was misleading. I edited the sentence.

>     4. Would it make sense to have a simple example of how to use the package in the paper?

Do you mean a code snippet like those mentioned in the [documentation](https://gym-saturation.readthedocs.io/en/latest/?badge=latest#how-to-use)?
