### 2021-11-16 #3849 @lutzhamel
@inpefess that was a quick turnaround! Here are some follow up thoughts,

> This package uses doctest and is 100% covered. The tests are accessible to a user as part of docstrings of functions and classes, like here

I am not familiar with 'doctest'.  Would it make sense to have sentence or two in the 'README' file pointing out this functionality and how to use it?

> Do you mean a code snippet like those mentioned in the documentation?

Yes, maybe an example of an agent trying to get a reward for proving a simple hypothesis.
