### 2021-11-16 a215941
mention how to run the tests for the package explicitly
