+++ README.md
@@ -41,8 +41,13 @@ poetry install
 pre-commit install
 ```
 
-To check the code quality before creating a pull request, one might run the script `show_report.sh`. It locally does nearly the same as the CI pipeline after the PR is created.
+All the tests in this package are [doctests](https://docs.python.org/3/library/doctest.html). One can run them with the following command:
 
+```sh
+pytest --doctest-modules gym-saturation
+```
+
+To check the code quality before creating a pull request, one might run the script `show_report.sh`. It locally does nearly the same as the CI pipeline after the PR is created.
 
 # Reporting issues or problems with the software
 
