### 2021-11-16 8bd48cb
add shorter example
