### 2021-11-16 b627ead
add usage example to the paper
