### 2021-11-16 #3849 @inpefess
@lutzhamel thanks for your reply.

> I am not familiar with 'doctest'. Would it make sense to have sentence or two in the 'README' file pointing out this functionality and how to use it?

Yes, that might be helpful for other users too. I added it to the `How to contribute` section.

> Yes, maybe an example of an agent trying to get a reward for proving a simple hypothesis.

I added a new, more concise example and included it in the paper.
