### 2021-11-17 ca090bd
add agent state to the base class
