### 2021-11-20 #3849 @lutzhamel
@inpefess @VivianePons That all looks good.  I have two minor additional comments:

1. Please also explain the acronym TPTP in the introductory paragraph
2. Does it make sense to have some references from the "Architecture" section to the "Usage" section?  It seems to me that Figure 1 is a precise embodiment of the usage example.

This concludes my review and I recommend publication.
