### 2021-11-20 9492f40
expand TPTP acronym
