### 2021-11-20 4ef1f9a
links from architecture section to the example
