### 2021-11-20 #3849 @inpefess
@lutzhamel Thank you! Your comments helped me to improve the draft quality.

>     1. Please also explain the acronym TPTP in the introductory paragraph

Done.

>     2. Does it make sense to have some references from the "Architecture" section to the "Usage" section?  It seems to me that Figure 1 is a precise embodiment of the usage example.

Yes, it looks like a good idea. I mentioned relevant lines of the example in the beginning  of the "Architecture" section.
