### 2021-12-07 30f48fe
mention Slurm in the documentation
