+++ doc/source/testing-an-agent.rst
@@ -21,6 +21,8 @@ Suppose you already have a trained agent implemented. Then you can use `an agent
 
   find $TPTP_HOME/Problems/*/*-*.p | parallel --bar --jobs 80% --timeout 30000% python agent_testing.py --problem_file {} --output_folder TPTP_CNF --step_limit 20
 
+Or you can use `Slurm Workload Manager <https://slurm.schedmd.com/>`__. See `an example <https://github.com/inpefess/gym-saturation/blob/master/gym_saturation/slurm-jobs>`__ from the project's repo.
+  
 You can write your own agent testing script based on ``agent_testing.py`` by calling ``episode`` function with your agent as an argument.
 
 After processing problems, you can get a report about its performance::
