+++ gym_saturation/envs/saturation_env.py
@@ -17,8 +17,10 @@ import os
 import random
 from typing import Dict, List, Tuple
 
-from gym import Env
+import numpy as np
+from gym import Env, spaces
 
+from gym_saturation.clause_space import ClauseSpace
 from gym_saturation.grammar import Clause
 from gym_saturation.logic_ops.factoring import all_possible_factors
 from gym_saturation.logic_ops.paramodulation import all_paramodulants_from_list
@@ -37,6 +39,7 @@ from gym_saturation.parsing.tptp_parser import TPTPParser, clause_to_tptp
 
 INFERRED_CLAUSES_PREFIX = "_"
 STATE_DIFF_UPDATED = "state_diff_updated"
+MAX_CLAUSES = 1000
 
 
 class SaturationEnv(Env):
@@ -53,14 +56,18 @@ class SaturationEnv(Env):
     >>> problem_list = sorted(glob(
     ...     os.path.join(tptp_folder, "Problems", "*", "*1-1.p")
     ... ))
-    >>> env = SaturationEnv(3, problem_list)
-
-    there is nothing non-deterministic here, but the seed can be set
-
+    >>> env = SaturationEnv(3, problem_list, 1)
     >>> env.seed(0)
     0
-
-    >>> len(env.reset())
+    >>> env.reset()
+    Traceback (most recent call last):
+     ...
+    ValueError: Too many clauses: 4
+    consider increasing `max_clauses` parameter of `__init__`
+    >>> env = SaturationEnv(3, problem_list)
+    >>> env.seed(0)
+    0
+    >>> len(env.reset()["real_obs"])
     4
 
     one can look at the current state in TPTP format
@@ -133,6 +140,7 @@ class SaturationEnv(Env):
         self,
         step_limit: int,
         problem_list: List[str],
+        max_clauses: int = MAX_CLAUSES,
     ):
         super().__init__()
         self.step_limit = step_limit
@@ -140,6 +148,13 @@ class SaturationEnv(Env):
         self._step_count = 0
         self._inference_count = 0
         self._state: List[Clause] = []
+        self.action_space = spaces.Discrete(max_clauses)
+        self.observation_space = spaces.Dict(
+            {
+                "action_mask": spaces.Box(0, 1, shape=(max_clauses,)),
+                "real_obs": ClauseSpace(),
+            }
+        )
 
     def _init_clauses(self):
         problem = random.choice(self.problem_list)
@@ -154,11 +169,10 @@ class SaturationEnv(Env):
             clause.processed = False
         return clauses
 
-    def reset(self) -> list:
+    def reset(self) -> dict:
         self._step_count = 0
         self._inference_count = 0
         self._state = reindex_variables(self._init_clauses(), "X")
-        self.action_space = list(range(len(self._state)))
         return self.state
 
     def _add_to_state(self, new_clauses: List[Clause]) -> None:
@@ -216,8 +230,8 @@ class SaturationEnv(Env):
             + [(action, clause_to_dict(self._state[action]))]
         )
 
-    def step(self, action: int) -> Tuple[list, float, bool, dict]:
-        if action not in self.action_space:
+    def step(self, action: int) -> Tuple[dict, float, bool, dict]:
+        if self._state[action].processed:
             raise ValueError(f"action {action} is not valid")
         self._step_count += 1
         if self._state[action].literals == []:
@@ -243,15 +257,17 @@ class SaturationEnv(Env):
             or self._step_count >= self.step_limit
         ):
             return self.state, 0.0, True, {STATE_DIFF_UPDATED: updated}
-        self.action_space = [
-            i for i, clause in enumerate(self._state) if not clause.processed
-        ]
[…]
