+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "gym-saturation"
-version = "0.1.7"
+version = "0.1.8"
 description = "An OpenAI Gym environment for saturation provers"
 authors = ["Boris Shminke <boris@shminke.ml>"]
 license = "Apache-2.0"
@@ -34,7 +34,7 @@ jedi = "*"
 rope = "*"
 black = "*"
 jupyterlab = "*"
-mypy = "*"
+mypy = "!=0.930"
 pycodestyle = "*"
 pylint = "*"
 pytest = "*"
@@ -55,5 +55,5 @@ profile = "black"
 src_paths = ["gym_saturation"]
 
 [build-system]
-requires = ["poetry-core>=1.1.11"]
+requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
