### 2021-12-23 deb2dee
temporary disable type checks because of mypy + numpy bug
