### 2021-12-25 40bed7c
fix a bug in reflexivity resolution (negated vs non negated)
