### 2021-12-25 706e517
make TPTP the default representation of for clauses
