+++ gym_saturation/parsing/tptp_parser.py
@@ -19,7 +19,7 @@ from typing import List
 
 from lark import Lark, Token
 
-from gym_saturation.grammar import Clause, Function, Term
+from gym_saturation.grammar import Clause
 from gym_saturation.parsing.cnf_parser import CNFParser
 
 if sys.version_info.major == 3 and sys.version_info.minor >= 9:
@@ -32,17 +32,25 @@ else:
 # pylint: disable=too-few-public-methods
 class TPTPParser:
     """
+    >>> from gym_saturation.grammar import (Literal, Predicate, Variable,
+    ...     Function)
+    >>> clause = Clause([Literal(True, Predicate("this_is_a_test_case", [Function("f", [Variable("X")])]))], inference_rule="resolution", inference_parents=["one", "two"], label="clause")
+    >>> TPTPParser().parse(str(clause), "") == [clause]
+    True
     >>> tptp_parser = TPTPParser()
     >>> tptp_text = (
     ...     files("gym_saturation")
     ...     .joinpath("resources/TPTP-mock/Problems/TST/TST001-1.p")
     ...     .read_text()
     ... )
-    >>> tptp_parser.parse(
+    >>> print("\\n".join(map(str, tptp_parser.parse(
     ...     tptp_text,
     ...     files("gym_saturation").joinpath("resources/TPTP-mock")
-    ... )
-    [Clause(literals=[Literal(negated=False, atom=Predicate(name='this_is_a_test_case', arguments=[Function(name='test_constant', arguments=[])]))], label='this_is_a_test_case_1', inference_parents=['one', 'two'], inference_rule='resolution', processed=None, birth_step=None), Clause(literals=[Literal(negated=True, atom=Predicate(name='this_is_a_test_case', arguments=[Function(name='test_constant', arguments=[])]))], label='this_is_a_test_case_2', inference_parents=None, inference_rule=None, processed=None, birth_step=None), Clause(literals=[Literal(negated=False, atom=Predicate(name='=', arguments=[Function(name='test_constant', arguments=[]), Variable(name='X')]))], label='test_axiom', inference_parents=None, inference_rule=None, processed=None, birth_step=None), Clause(literals=[Literal(negated=True, atom=Predicate(name='=', arguments=[Function(name='test_constant', arguments=[]), Function(name='0', arguments=[])]))], label='test_axiom_2', inference_parents=None, inference_rule=None, processed=None, birth_step=None)]
+    ... ))))
+    cnf(this_is_a_test_case_1, hypothesis, this_is_a_test_case(test_constant), inference(resolution, [], [one, two])).
+    cnf(this_is_a_test_case_2, hypothesis, ~this_is_a_test_case(test_constant)).
+    cnf(test_axiom, hypothesis, =(test_constant, X)).
+    cnf(test_axiom_2, hypothesis, ~=(test_constant, 0)).
     """
 
     def __init__(self):
@@ -79,47 +87,3 @@ class TPTPParser:
                         self.parse(included_file.read(), tptp_folder)
                     )
         return clauses
-
-
-def _term_to_tptp(term: Term) -> str:
-    if isinstance(term, Function):
-        arguments = [_term_to_tptp(argument) for argument in term.arguments]
-        if arguments != []:
-            return f"{term.name}({','.join(arguments)})"
-    return term.name
-
-
-def clause_to_tptp(clause: Clause) -> str:
-    """
-    >>> from gym_saturation.grammar import Literal, Predicate, Variable
-    >>> clause = Clause([Literal(True, Predicate("this_is_a_test_case", [Function("f", [Variable("X")])]))], inference_rule="resolution", inference_parents=["one", "two"], label="clause")
-    >>> TPTPParser().parse(clause_to_tptp(clause), "") == [clause]
-    True
-
[…]
