+++ gym_saturation/agent_testing.py
@@ -13,20 +13,17 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-import json
-import os
 import random
 import sys
 from argparse import ArgumentParser, Namespace
 from dataclasses import dataclass
 from operator import itemgetter
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Optional
 
 import gym
 
 from gym_saturation.envs import SaturationEnv
 from gym_saturation.envs.saturation_env import STATE_DIFF_UPDATED
-from gym_saturation.grammar import Clause
 from gym_saturation.logic_ops.utils import clause_length
 from gym_saturation.parsing.json_grammar import dict_to_clause
 
@@ -71,33 +68,6 @@ class Transition:
     reward: float
     done: bool
     info: Dict[str, Any]
-
-
-def save_final_state(
-    problem_filename: str, output_folder: str, episode_memory: List[Transition]
-) -> None:
-    """
-    save the final environment state of solving a TPTP problem
-
-    :param problem_filename: a full path to the problem file
-    :param output_folder: where to write subfolders and JSON files
-    :param episode_memory: a list of transitions during the episode
-    :returns:
-    """
-    problem_name = os.path.splitext(os.path.basename(problem_filename))[0]
-    output_subfolder = os.path.join(
-        output_folder, os.path.basename(os.path.dirname(problem_filename))
-    )
-    try:
-        os.mkdir(output_subfolder)
-    except FileExistsError:
-        pass
-    with open(
-        os.path.join(output_subfolder, f"{problem_name}.json"),
-        "w",
-        encoding="utf-8",
-    ) as data_file:
-        data_file.write(json.dumps(episode_memory[-1].observation["real_obs"]))
 
 
 class SizeAgent(BaseAgent):
@@ -191,6 +161,7 @@ def episode(env: SaturationEnv, agent: BaseAgent) -> List[Transition]:
     """
     tries to solve the problem and logs the clauses
 
+    >>> import os
     >>> import shutil
     >>> test_agent_output = "test_agent_output"
     >>> shutil.rmtree(test_agent_output, ignore_errors=True)
@@ -206,22 +177,20 @@ def episode(env: SaturationEnv, agent: BaseAgent) -> List[Transition]:
     ...     .joinpath("resources/TPTP-mock/Problems/TST")
     ... , "*-*.p")))
     >>> random.seed(0)
-    >>> agents = [SizeAgent(), RandomAgent(), AgeAgent()]
-    >>> for i in range(3):
+    >>> agents = [SizeAgeAgent(2, 1), SizeAgeAgent(1, 2), RandomAgent()]
+    >>> for i in range(3):  # doctest: +ELLIPSIS
     ...     env = gym.make(
     ...         "gym_saturation:saturation-v0",
     ...         step_limit=5,
     ...         problem_list=[problem_list[i]],
     ...     )
-    ...     save_final_state(
-    ...         problem_list[i],
-    ...         test_agent_output,
-    ...         episode(env, agents[i])
-    ...     )
-    >>> print(sorted(agent_testing_report(
-    ...     problem_list + ["this_is_a_test_case"], test_agent_output
-    ... ).items()))
-    [('TST001-1', ('PROOF_FOUND', 2, 2)), ('TST002-1', ('STEP_LIMIT', 5, -1)), ('TST003-1', ('STEP_LIMIT', 4, -1)), ('this_is_a_test_case', ('ERROR', -1, -1))]
+    ...     agent_testing_report(env, agents[i])
+    Proof of length 1 found in 4 steps:
[…]
