### 2021-12-26 e1114b4
we don't output folder argument anymore
