### 2021-12-26 f9a363a
don't store episode memory in RAM
