### 2021-12-27 c004629
no duplicates in a proof
