### 2021-12-27 c9941ef
equality can be infix only in TPTP
