+++ gym_saturation/logic_ops/paramodulation.py
@@ -208,9 +208,9 @@ def all_paramodulants_from_list(
     >>> res = all_paramodulants_from_list([one], two, "inferred_", 0)
     >>> dedup = map(Clause, deduplicate([clause.literals for clause in res]))
     >>> print("\\n".join(map(str, dedup)))
-    cnf(None, hypothesis, =(a, c) | =(X, X)).
-    cnf(None, hypothesis, =(c, b) | =(a, b)).
-    cnf(None, hypothesis, =(b, c) | =(a, b)).
+    cnf(None, hypothesis, a = c | X = X).
+    cnf(None, hypothesis, c = b | a = b).
+    cnf(None, hypothesis, b = c | a = b).
 
     :param clauses: a list of (processed) clauses
     :param given_clause: a new clause which should be combined with all the
