### 2021-12-27 912f639
clause label can't start with an underscore in TPTP
