+++ gym_saturation/envs/saturation_env.py
@@ -37,7 +37,7 @@ from gym_saturation.logic_ops.utils import (
 from gym_saturation.parsing.json_grammar import clause_to_dict
 from gym_saturation.parsing.tptp_parser import TPTPParser
 
-INFERRED_CLAUSES_PREFIX = "_"
+INFERRED_CLAUSES_PREFIX = "i"
 STATE_DIFF_UPDATED = "state_diff_updated"
 MAX_CLAUSES = 100000
 
@@ -121,7 +121,7 @@ class SaturationEnv(Env):
     for validation purposes)
 
     >>> print(env.tstp_proof)
-    cnf(_0, hypothesis, $false, inference(resolution, [], [this_is_a_test_case_1, this_is_a_test_case_2])).
+    cnf(i0, hypothesis, $false, inference(resolution, [], [this_is_a_test_case_1, this_is_a_test_case_2])).
 
     >>> env = SaturationEnv(1, problem_list)
 
