+++ gym_saturation/grammar.py
@@ -13,8 +13,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from typing import List, Optional, Union
+from uuid import uuid1
 
 
 @dataclass
@@ -101,7 +102,7 @@ class Clause:
     """
 
     literals: List[Literal]
-    label: Optional[str] = None
+    label: str = field(default_factory=lambda: str(uuid1()))
     inference_parents: Optional[List[str]] = None
     inference_rule: Optional[str] = None
     processed: Optional[bool] = None
