### 2021-12-27 6bfcb9d
fix another bug in reflexivity resolution
