### 2021-12-27 371224f
add all examples to Binder
