### 2021-12-27 d9b1892
TPTP labels can't start with numbers and contain dashes
