+++ gym_saturation/__init__.py
@@ -18,4 +18,6 @@ from gym.envs.registration import register
 register(
     id="saturation-v0",
     entry_point="gym_saturation.envs:SaturationEnv",
+    max_episode_steps=1000,
+    reward_threshold=1.0,
 )
