### 2021-12-28 00c2748
use automodule whenever possible
