+++ gym_saturation/grammar.py
@@ -1,17 +1,19 @@
+# Copyright 2021 Boris Shminke
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     https://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 """
-Copyright 2021 Boris Shminke
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    https://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+Grammar
+********
 """
 from dataclasses import dataclass, field
 from typing import List, Optional, Union
@@ -20,14 +22,22 @@ from uuid import uuid1
 
 @dataclass
 class Variable:
-    """ a variable is characterised only by its name """
+    """
+    .. _variable:
+
+    a variable is characterised only by its name
+    """
 
     name: str
 
 
 @dataclass
 class Function:
-    """ a functional symbol might be applied to a list of arguments """
+    """
+    .. _Function:
+
+    a functional symbol might be applied to a list of arguments
+    """
 
     name: str
     arguments: List[Union[Variable, "Function"]]
@@ -38,7 +48,11 @@ Term = Union[Variable, Function]
 
 @dataclass
 class Predicate:
-    """ a predicate symbol might be applied to a list of arguments """
+    """
+    .. _Predicate:
+
+    a predicate symbol might be applied to a list of arguments
+    """
 
     name: str
     arguments: List[Term]
@@ -49,7 +63,11 @@ Proposition = Union[Predicate, Term]
 
 @dataclass
 class Literal:
-    """ literal is an atom which can be negated or not """
+    """
+    .. _Literal:
+
+    literal is an atom which can be negated or not
+    """
 
     negated: bool
     atom: Predicate
@@ -86,6 +104,8 @@ def _literal_to_tptp(literal: Literal) -> str:
 @dataclass
 class Clause:
     """
+    .. _Clause:
+
     clause is a disjunction of literals
 
     :param literals: a list of literals, forming the clause
