### 2022-01-16 9703383
drop Python 3.6 support, add Python 3.10 support
