+++ gym_saturation/envs/saturation_env.py
@@ -1,4 +1,4 @@
-# Copyright 2021 Boris Shminke
+# Copyright 2021-2022 Boris Shminke
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
