+++ gym_saturation/utils.py
@@ -1,4 +1,4 @@
-# Copyright 2021 Boris Shminke
+# Copyright 2021-2022 Boris Shminke
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -36,5 +36,5 @@ def deduplicate(a_list: List[Any]) -> List[Any]:
 
 
 def pickle_copy(obj: object):
-    """ faster ``deepcopy`` analogue with ``pickle`` """
+    """faster ``deepcopy`` analogue with ``pickle``"""
     return pickle.loads(pickle.dumps(obj))
