+++ README.rst
@@ -73,7 +73,7 @@ welcome. To start:
 
    git clone https://github.com/inpefess/gym-saturation
    cd gym-saturation
-   # activate python virtual environment with Python 3.6+
+   # activate python virtual environment with Python 3.7+
    pip install -U pip
    pip install -U setuptools wheel poetry
    poetry install
