+++ gym_saturation/parsing/tptp_parser.py
@@ -54,8 +54,8 @@ class TPTPParser:
     ... ))))
     cnf(this_is_a_test_case_1, hypothesis, this_is_a_test_case(test_constant), inference(resolution, [], [one, two])).
     cnf(this_is_a_test_case_2, hypothesis, ~this_is_a_test_case(test_constant)).
-    cnf(test_axiom, hypothesis, test_constant = X).
-    cnf(test_axiom_2, hypothesis, ~test_constant = 0).
+    cnf(test_axiom, axiom, test_constant = X).
+    cnf(test_axiom_2, axiom, ~test_constant = 0).
     """
 
     def __init__(self):
