+++ doc/source/what-is-going-on.rst
@@ -39,7 +39,7 @@ The environment's state is a list of logical clauses. Each clause is a list of l
 
 Literal is a predicate, negated or not. A predicate can have arguments, which can be functions or variables. Functions can have arguments, which in turn can be functions or variables.
 
-Grammar is encoded in Python objects in a self-explanatory way. Each grammar object is a dictionary with an obligatory key ``class`` (:ref:`Clause <Clause>`, :ref:`Literal <Literal>`, :ref:`Predicate <Predicate>`, :ref:`Function <Function>`, :ref:`Variable <Variable>`), and other keys representing this object's properties (such as being negated or having a list of arguments). To parse these JSON representation into package's inner representation, use :ref:`dict_to_clause <dict_to_clause>`.
+Grammar is encoded in Python objects in a self-explanatory way. Each grammar object is a dictionary with an obligatory key ``class`` (:ref:`Clause <Clause>`, :ref:`Literal <Literal>`, :ref:`Predicate <Predicate>`, :ref:`Function <Function>`, :ref:`Variable <Variable>`), and other keys representing this object's properties (such as being negated or having a list of arguments).
 
 What is an Observation
 ***********************
