### 2022-02-08 1cdcad0
immutable but typed NamedTuples instead of mutable dataclasses
