### 2022-02-08 d2d0395
asdict is too slow
