+++ gym_saturation/parsing/cnf_parser.py
@@ -67,8 +67,8 @@ class CNFParser(Transformer):
         a functional symbol with arguments
         """
         if len(children) > 1:
-            return Function(children[0], children[1])
-        return Function(children[0], [])
+            return Function(children[0], tuple(children[1]))
+        return Function(children[0], ())
 
     def fof_defined_plain_formula(self, children):
         """
@@ -104,12 +104,12 @@ class CNFParser(Transformer):
 
         a list of arguments, organised in pairs
         """
-        result = []
+        result = ()
         for item in children:
-            if isinstance(item, list):
-                result.extend(item)
+            if not isinstance(item, (Variable, Function)):
+                result = result + tuple(item)
             else:
-                result.append(item)
+                result = result + (item,)
         return result
 
     @staticmethod
@@ -133,8 +133,8 @@ class CNFParser(Transformer):
     def _predicate(children):
         """predicates are atomic formulae"""
         if len(children) > 1:
-            return Predicate(children[0], children[1])
-        return Predicate(children[0], [])
+            return Predicate(children[0], tuple(children[1]))
+        return Predicate(children[0], ())
 
     def fof_plain_atomic_formula(self, children):
         """
@@ -149,7 +149,7 @@ class CNFParser(Transformer):
 
         some predicates are in the infix form, so we translate to them prefix
         """
-        return Predicate(children[1], [children[0], children[2]])
+        return Predicate(children[1], (children[0], children[2]))
 
     @staticmethod
     def fof_infix_unary(children):
@@ -158,7 +158,7 @@ class CNFParser(Transformer):
 
         some predicates are in the infix form, so we translate to them prefix
         """
-        return Predicate(children[1], [children[0], children[2]])
+        return Predicate(children[1], (children[0], children[2]))
 
     @staticmethod
     def disjunction(children):
@@ -168,26 +168,26 @@ class CNFParser(Transformer):
         basic clause structure
         """
         if len(children) == 1:
-            return Clause(children)
-        literals = []
-        for item in [children[0], children[2]]:
+            return Clause(tuple(children))
+        literals = ()
+        for item in (children[0], children[2]):
             if isinstance(item, Clause):
-                literals.extend(item.literals)
+                literals = literals + item.literals
             else:
-                literals.append(item)
+                literals = literals + (item,)
         return Clause(literals)
 
     @staticmethod
     def _parse_inference_parents(inference_parents):
         if isinstance(inference_parents[0], list):
-            clause_inference_parents = list(
+            clause_inference_parents = tuple(
                 map(
                     itemgetter(0),
                     inference_parents,
                 )
             )
         else:
-            clause_inference_parents = [inference_parents[0]]
+            clause_inference_parents = (inference_parents[0],)
         return clause_inference_parents
 
     def cnf_annotated(self, children):
@@ -229,7 +229,7 @@ class CNFParser(Transformer):
         """
         if isinstance(children, list):
             if len(children) == 1:
-                return children[0]
+                return tuple(children[0])
         return children
 
     @staticmethod
