+++ mypy.ini
@@ -1,6 +1,6 @@
 [mypy]
 show_error_codes = True
-disable_error_code = no-redef,attr-defined,name-defined
+disable_error_code = no-redef
 
 [mypy-gym]
 ignore_missing_imports = True

[…]
