### 2022-02-09 ba6676b
orjson + dataclasses
