### 2022-02-11 2caa012
don't fail when reaching ``max_clause`` but finish the episode
