### 2022-02-11 b6239b7
return positive actions during the last step
