### 2022-02-25 #3849 @quickbeam123
"In contrast, for interactive theorem provers, some projects like
HOList (Bansal, Loos, Rabe, Szegedy, & Wilcox, 2019)
separate the concepts of environment and agent.
Such modular architecture proved to help create multiple successful agents by
different groups of researchers (see, e.g. Paliwal et al. (2020) or Bansal, Loos, Rabe, &
Szegedy (2019))."

- While I agree in principle, the example is misleading in the sense that both
the papers cited by the second sentence are by the same team as the HOList paper itself.
Could you find a different example? While I did not explore this in detail myself
(and agree that HOList is marketed as an environment / gym for agent training and experimentation),
it could well be that experiments with HOList are not so easy without the corresponding Google
infrastructure available exclusively to Szegedy et al.

"to be refutation complete"
- typically, one says "refutationally complete"

"L[t] is a result of substituting the term r in L[r] for the term t at only one chosen position."
- I would think "... substituting the term t in L[r] for the term r ..."

- "We represent the clauses Python classes" a missing word?
