+++ joss-paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 # Statement of need
 
-Current applications of RL to saturation-based ATPs like Enigma [@DBLP:conf/cade/JakubuvCOP0U20] or Deepire [@DBLP:conf/cade/000121a] are similar in that the environment and the agent are not separate pieces of software but parts of larger systems that are hard to disentangle. The same is true for non saturation-based RL-friendly provers too (e.g. lazyCoP, @DBLP:conf/tableaux/RawsonR21). This monolithic approach hinders free experimentation with novel machine learning (ML) models and RL algorithms and creates unnecessary complications for ML and RL experts willing to contribute to the field. In contrast, for interactive theorem provers, some projects like HOList [@DBLP:conf/icml/BansalLRSW19] separate the concepts of environment and agent. Such modular architecture proved to help create multiple successful agents by different groups of researchers (see, e.g. @DBLP:conf/aaai/PaliwalLRBS20 or @DBLP:journals/corr/abs-1905-10501). `gym-saturation` is an attempt to implement a modular environment-agent architecture of an RL-based ATP. In addition, some RL empowered saturation ATPs are not accompanied with their source code [@DBLP:journals/corr/abs-2106-03906], while `gym-saturation` is open-source software.
+Current applications of RL to saturation-based ATPs like Enigma [@DBLP:conf/cade/JakubuvCOP0U20] or Deepire [@DBLP:conf/cade/000121a] are similar in that the environment and the agent are not separate pieces of software but parts of larger systems that are hard to disentangle. The same is true for non saturation-based RL-friendly provers too (e.g. lazyCoP, @DBLP:conf/tableaux/RawsonR21). This monolithic approach hinders free experimentation with novel machine learning (ML) models and RL algorithms and creates unnecessary complications for ML and RL experts willing to contribute to the field. In contrast, for interactive theorem provers, projects like HOList [@DBLP:conf/icml/BansalLRSW19] or GamePad [DBLP:conf/iclr/HuangDSS19] separate the concepts of environment and agent. In contrast, for interactive theorem provers, projects like HOList [@DBLP:conf/icml/BansalLRSW19] or GamePad [DBLP:conf/iclr/HuangDSS19] separate the concepts of environment and agent. Such modular architecture may lead to the development of easily comparable agents based on diverse approaches (see, e.g. @DBLP:conf/aaai/PaliwalLRBS20 or @DBLP:journals/corr/abs-1905-10501). `gym-saturation` is an attempt to implement a modular environment-agent architecture of an RL-based ATP. In addition, some RL empowered saturation ATPs are not accompanied with their source code [@DBLP:journals/corr/abs-2106-03906], while `gym-saturation` is open-source software.
 
 # Usage example
 
