+++ joss-paper/paper.bib
@@ -66,8 +66,6 @@
   biburl    = {https://dblp.org/rec/conf/cade/0001CV19.bib},
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
-
-
 @inproceedings{DBLP:conf/cav/KovacsV13,
   author    = {Laura Kov{\'{a}}cs and
                Andrei Voronkov},
@@ -259,5 +257,20 @@ eprint = {https://doi.org/10.1137/0204036}
   eprint    = {1905.10501},
   timestamp = {Mon, 03 Jun 2019 13:42:33 +0200},
   biburl    = {https://dblp.org/rec/journals/corr/abs-1905-10501.bib},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+@inproceedings{DBLP:conf/iclr/HuangDSS19,
+  author    = {Daniel Huang and
+               Prafulla Dhariwal and
+               Dawn Song and
+               Ilya Sutskever},
+  title     = {GamePad: {A} Learning Environment for Theorem Proving},
+  booktitle = {7th International Conference on Learning Representations, {ICLR} 2019,
+               New Orleans, LA, USA, May 6-9, 2019},
+  publisher = {OpenReview.net},
+  year      = {2019},
+  url       = {https://openreview.net/forum?id=r1xwKoR9Y7},
+  timestamp = {Thu, 25 Jul 2019 14:25:40 +0200},
+  biburl    = {https://dblp.org/rec/conf/iclr/HuangDSS19.bib},
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
