+++ joss-paper/paper.bib
@@ -147,29 +147,15 @@
   biburl    = {https://dblp.org/rec/conf/tableaux/RawsonR21.bib},
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
-@article{DBLP:journals/corr/abs-2106-03906,
-  author    = {Ibrahim Abdelaziz and
-               Maxwell Crouse and
-               Bassem Makni and
-               Vernon Austil and
-               Cristina Cornelio and
-               Shajith Ikbal and
-               Pavan Kapanipathi and
-               Ndivhuwo Makondo and
-               Kavitha Srinivas and
-               Michael Witbrock and
-               Achille Fokoue},
-  title     = {Learning to Guide a Saturation-Based Theorem Prover},
-  journal   = {CoRR},
-  volume    = {abs/2106.03906},
-  year      = {2021},
-  url       = {https://arxiv.org/abs/2106.03906},
-  eprinttype = {arXiv},
-  eprint    = {2106.03906},
-  timestamp = {Thu, 10 Jun 2021 16:34:18 +0200},
-  biburl    = {https://dblp.org/rec/journals/corr/abs-2106-03906.bib},
-  bibsource = {dblp computer science bibliography, https://dblp.org}
-}
+@ARTICLE{9669114,
+author={Abdelaziz, Ibrahim and Crouse, Maxwell and Makni, Bassem and Austel, Vernon and Cornelio, Cristina and Ikbal, Shajith and Kapanipathi, Pavan and Makondo, Ndivhuwo and Srinivas, Kavitha and Witbrock, Michael and Fokoue, Achille},
+journal={IEEE Transactions on Pattern Analysis and Machine Intelligence}, 
+title={Learning to Guide a Saturation-Based Theorem Prover}, 
+year={2022},
+volume={},
+number={},
+pages={1-1},
+doi={10.1109/TPAMI.2022.3140382}}
 @inproceedings{DBLP:conf/icml/BansalLRSW19,
   author    = {Kshitij Bansal and
                Sarah M. Loos and
