### 2022-03-02 #3849 @VivianePons
Great! @inpefess can you please make a tagged release and archive (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or other). Please make sure that the title match the paper title and same for authors.
Once done, please send me the link / doi so that I can update the info and move to paper acceptance
