### 2022-03-02 #3849 @inpefess
@VivianePons here is the release: https://github.com/inpefess/gym-saturation/releases/tag/v0.2.9
and here is the DOI from Zenodo: https://doi.org/10.5281/zenodo.6324282
